gcc warning for 32bit
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/loggertools/trunk@57 fae35197-53e2-0310-b5ce-cc51903860e9

+++ src/filser.c
@@ -616,7 +616,7 @@ static int download_flight(int argpos, int argc, char **argv) {
             break;
 
         printf("downloading section %u, %lu bytes\n",
-               z, section_lengths[z]);
+               z, (unsigned long)section_lengths[z]);
 
         ret = download_section(fd, z,
                                p, section_lengths[z]);
